+++ package.json
@@ -1,0 +1,18 @@
+{
+  "name": "alchemy-cd-integration-test",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "start": "mocha --reporter=spec"
+  },
+  "dependencies": {
+    "mocha": "*",
+    "should": "*",
+    "btoa": "*",
+    "nconf":  "*",
+    "validator": "*",
+    "jquery":"*",
+    "request" : "*",
+    "request-json":"*"
+  }
+}
